Refactor: rename package names
